exchange/requirements.txt
alembic==1.4.3
amqp==5.0.5
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.5
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dj-database-url==0.5.0
Django==3.1.7
django-celery-beat==2.2.0
django-rest-framework==0.1.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.0.2
lazy-object-proxy==1.4.3
lxml==4.6.2
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
prompt-toolkit==3.0.17
psycopg2==2.8.6
psycopg2-binary==2.8.6
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
python-crontab==2.5.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
redis==3.5.3
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
